requirements/requirements.txt
django
djangorestframework
drf-yasg
django-cors-headers
mysqlclient
django-mysql
redis
celery
flake8